requirements.txt
folium==0.12.1
haversine==2.3.1
h5py==2.10.0
matplotlib==3.4.2
networkx==2.5.1
numpy
pandas
tensorflow-gpu==1.15.0
pydot-ng==2.0.0
graphviz==0.16
impyla